test(skills): pin the four boundary fixes, and split the case module at its seam
Six curated cases for the four corruptions the previous commit closed: an
indented code block ending a pending definition, a list marker ending one in
each of its two states, a dedented title NOT closing the item holding its
definition, and a whitespace-only label. Plus the control that whitespace
INSIDE a real label still leaves it a definition.

Each fix is mutation-checked separately, and each mutation fails three tests:
the oracle comparison in both suites and the public repair no-op. Reverting
the lazy-continuation guard, the list-item clear, the indented-code clear, or
the blank-label class all behave the same way.

The added cases took commonmark_fence_cases.py to 517 lines and the taste
ratchet to 576 against a baseline of 575. Raising a baseline to clear a gate
is forbidden here, so the module is split along the seam it already had:
curated cases and the oracle stay, the generator and its baselines move to
commonmark_fence_fuzz.py. 413 and 116 lines; the ratchet is back at 575.

The moved comment block referred to "the curated cases above" and to "this
module's docstring". Both were true where it used to sit and false where it
now sits, so both are repointed rather than left to read plausibly and mean
nothing.
